core: OsinfoPropertiesParserTest failure
When the OsinfoPropertiesParser fails to parse a file it throws a
pretty readable exception.

This patch clears up the test to just allow this exception to be
raised outside the test method and error it out instead of
boiler-plate catching it and failing the test.
As a side bonus, this means that System.err isn't explicitly used
anymore, making the test friendlier to CI environments and shell
output redirections.

Example of a failed test produced by manually corrupting
packaging/conf/osinfo-defaults.properties:

Tests in error:
  defaultProperties(org.ovirt.engine.core.utils.osinfo.OsinfoPropertiesParserTest): osinfo properties syntax error: /home/amureini/src/git/ovirt-engine/backend/manager/modules/utils/../../../../packaging/conf/osinfo-defaults.properties: 36, 20: error: Encountered " <OS_UNIQUE_NAME> "0a "" at line 36, column 21.(..)

Change-Id: If828dda51e4ad7fecd3a6f6e38d9aa4933fcaf65
Signed-off-by: Allon Mureinik <amureini@redhat.com>