core: Consolidate Transfer[Disk]Image[Command]
TransferImageCommand is an abstract class with just one concrete
implementer, TransferDiskImageCommand.
It's unlikely to eve be used on its own, and when such a usecase
arises, it can always be split out. In the current implementation, it
serves no purpose, and just confuses the reader by scattering the
implementation among several classes.

This patch consolidates the following:
- TransferImageCommand in to TransferDiskImageCommand
- TransferImageParameters in to TransferDiskImageParameters

protected methods inherited from the base class were converted to
private in the process, as there is no longer an extending class that
depends on them.

Change-Id: I17bc04f4a70bf2401dfd7c726dce1e25d822ded0
Signed-off-by: Allon Mureinik <amureini@redhat.com>