core: TransferDiskImageCommand stream closing
In addImageTicketToProxy, the stream is handled with the
try-with-resource syntax introduced in Java 7, which closes the
stream when it exits the try's body.
Explicitly closing the stream in the try's body is redundant at best,
and just plain wrong at worst.

This patch removes the explicit close() call and allows Java's syntax
to do the lifting for us.

Change-Id: I3e286ccc82e35728eafa96ace732fcb56510d7b2
Signed-off-by: Allon Mureinik <amureini@redhat.com>