core: verify lease storage domain is active before removing a VM
Currently, there is no validation whether the storage domain that contains
the VM lease is active, therefore, when a VM with a lease is removed,
the VM managed to remove but the lease remains on the domain.

This patch adds a validation to verify that the lease storage domain is
active, as similar to the VM disks storage domain validation.

Change-Id: Ibd74c21dabb5a4c4b64a3812c9b3ec3d69305beb
Bug-Url: https://bugzilla.redhat.com/1561447
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>
(cherry picked from commit a3ea4c2447e8616acf424120ecdc4f9344b50588)