core: Fix thread pool monitoring logs
We used to log:

    Thread pool 'Foo' is using 0 threads out of 5 and 2 tasks are waiting in the queue.

Which looks like a thread pool without any threads, that will never make
any progress processing the tasks in the queue. But here is what this
log show:

- 0 threads in state NEW, RUNNABLE, or BLOCKED
- 2 threads in state TIMED_WAITING or WAITING
- number of tasks in queue is unknown

For "engine" thread pool, we log more clear message:

    Thread pool 'engine' is using 0 threads out of 100, 8 threads waiting for tasks and 0 tasks in queue.

In this case the last number is the number of tasks in the queue.

Change the text to describe the actual values, matching 'engine' thread
pool log:

    Thread pool 'Foo' is using 0 threads out of 5, 2 threads are waiting for tasks.

This message is not correct yet since threads in WAITING or
TIMED_WAITING state may not wait for tasks. For example they may be
running commands that use Object.wait() to signal completion.

Change-Id: Ia9214628339e09b4e7b8c247d51a968d0e17a629
Related-To: https://bugzilla.redhat.com/1556971
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
(cherry picked from commit b1b6cc470e4865d6568ba677cf3fb4795f096f82)