core: Introduce ExecutorServiceRule
ThreadPoolUtil relies on a static ExecutorService member in order to
create and manage its executor thread. This member is explicitly
initialized in the engine via the EngineThreadPools bean's init()
method.

For most unit tests, it's initialized by BaseCommandTest's @Before
method, which poses several problems:
1. It requires tests that use (even implicitly)  ThreadPoolUtil to
   extend BaseCommandTest.
2. The initialized service is never cleared, making the entire test
   suite non-deterministic. If a test that extends BaseCommandTest
   happens to have run before a certain test X, X will have an
   initialized ThreadPoolUtil, and if it hasn't, the ThreadPoolUtil
   won't be initialized. This has lead, e.g., to observed
   non-deterministic NullPointerExceptions in
   StorageDeviceSyncJobTest (see, e.g., [1]).

This patch fixes the aforementioned issue by introducing a new
ExecutorServiceRule that sets up the ThreadPoolUtil at the test's
start and resets it at the test's end, and adds it as a @ClassRule to
the tests that rely on the ThreadPoolUtil.

[1] http://jenkins.ovirt.org/job/ovirt-engine_4.2_check-patch-el7-x86_64/542/

Change-Id: I0bba6b2e7a98867c0d6339ecda3a448a06455deb
Signed-off-by: Allon Mureinik <amureini@redhat.com>