db: change LUN ID parameter type
Since the luns table declares the LUN ID as type "character
varying(255)", and the disk_lun_map table declares it as type "character
varying", there's no reason to limit the size of the input parameter
"v_LUN_id" to 50 characters.
Therefore, this patch changes its size to 255 in all occurrences.

Change-Id: I591186c5cd7824adc114ef68705b1b493bd1b0ed
Signed-off-by: Idan Shaby <ishaby@redhat.com>