engine: Ignore failure connecting ISO and Export SD
When adding a new host or activating an existing one, if the
host have no access to an existing ISO or Export storage
domain it will never get to Up state.

In this patch, the call to the Vdsm to connect File storage domains
is split, so that the Data SD will not be sent in the same request
as ISO/import-export SDs.
In addition, the result of the connections of these type of SDs is
ignored.

Also, in order to prevent to trigger an event to VDSEvenListener, when
getting a NetworkException when connecting to ISO/Export-Import, an
additional flag is added to the parameters of ConnectStorageServer.

Change-Id: I4d73f42d866ce024ba0282e1186b8c3a02b12098
Bug-Url: https://bugzilla.redhat.com/1499056
Signed-off-by: Fred Rolland <frolland@redhat.com>