core: Improve ObjectDescriptor
Originally, ObjectDescriptor was used to print compmlex types in a
flat way. It supported objects of types Map and Map[]. Howevrer, if
the maps contained values of type Object[], the array members weren't
printed, but rather default toString of the array, instead of a deep
toString implementation. In this patch, a single generic
toStringBuilder API is introduced to support all types. In addition
the printings changed to support standrad format: {} for maps and []
for oterable objects.

Change-Id: Idc3e373b1c50b78ac3862573c1d0eb14409c7262
Signed-off-by: Ala Hino <ahino@redhat.com>