db: enlarge variable size to avoid SQL exception
The luns table declares the LUN ID as type character varying(255).
Still, the function DeleteLUN declares the temporary variable v_val as
VARCHAR(50), and tries to store the LUN ID in it.
As a consequence, in case that a LUN with an ID of more than 50
characters is deleted, an SQL exception occurs because the value is too
long to be stored in this variable.
Therefore, this patch changes this variable size to be 255 rather than
50 in order to fix this bug.

Change-Id: Iad54d4f6c730fe3f195ea0aa492f9beac6bfe14e
Bug-Url: https://bugzilla.redhat.com/1565173
Signed-off-by: Idan Shaby <ishaby@redhat.com>