db: remove unneeded locking mechanism in DeleteLUN
The DeleteLUN procedure maintains a locking mechanism using a shared
variable, v_val, to avoid changing the LUN row while trying to delete
it.
Since in PostgreSQL, the FOR UPDATE lock is acquired by the DELETE
statement on a row [1], this locking mechanism is not needed and thus
removed in this patch.

[1] https://www.postgresql.org/docs/9.5/static/explicit-locking.html#LOCKING-ROWS

Change-Id: I15c2f119dc1257342ef6b1cdc14f372a3a973eeb
Signed-off-by: Idan Shaby <ishaby@redhat.com>