core: consume file-based volumes deferred preallocation
Currently, the engine order to VDSM to copy/move preallocated
file disk and create it as sparse on the new storage domain.

After patch will be merged-
https://gerrit.ovirt.org/#/c/93557
When copying/moving a preallocated volume, caller can specify
initialSize for file based volumes.
This will preallocate only an initialSize of bytes and create
an invalid raw volume.
The volume will be fully preallocated later when copying the data
into the new volume.
VDSM will keep file disk as preallocated after cold storage
migration and when cloning a VM from template.
So there is no need for the engine to manipulate the file to be
sparse instead of preallocated volume.

This patch fix the above behavior.

Change-Id: I824b21dfd45ead7d229aebf00b89280a47cc8fa8
Bug-Url: https://bugzilla.redhat.com/1571285
Bug-Url: https://bugzilla.redhat.com/1403183
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>