core: When balancing, subtract VM memory and CPU load from current host
During balancing and affinity enforcement, the VM is scheduled
as if it was not running. The host where it is running is more
loaded and so the VM may be scheduled elsewhere, even if the current host
is the best.
This can cause a VM to migrate between hosts in a loop and never stop.

Change-Id: I71569b6e26283da346092e5d4f8a2b17b4183d55
Bug-Url: https://bugzilla.redhat.com/1535175
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>