core: Standardize VnicProfileValidator's dependencies
VnicProfileValidator and doesn't have a reasonable agreed way to
manage its dependencies - some are passed via constructors and some
are used via the legacy getInstance() method.

This patch cleans up this class by settling on a single method
for all dependencies. Since the de-facto standard and least intrusive
way was to use the Injector, this method was selected and then
applied to all the class' dependencies.

Change-Id: I5585d63759ed7843e0ec3bc6543e0191c52727b6
Signed-off-by: Allon Mureinik <amureini@redhat.com>