core: Standardize NetwrokValidator's dependencies
NetworkValidator and its children don't have a reasonable agreed way
to manage their dependencies - some are passed via constructors, some
injected by using the Injector and some via the legacy getInstance()
methods.

This patch cleans up this hierarchy by settling on a single method
for all dependencies. Since the most common and least intrusive way
was to use the Injector, this method was selected and then applied to
all the classes' dependencies.

Change-Id: I2b8e8b826e6b666aecf1abff6a22dfbe2f19532d
Signed-off-by: Allon Mureinik <amureini@redhat.com>