core: Remove DbFacade.getInstance()
With most of the calls to DbFacade.getInstance() already removed,
this patch finishes the job by replacing the several dozen calls that
could not be replaced more elegantly with calls to Injector.get and
finally removes this deprecated method.

Unit tests were changed to use the InjectorRule to mock such
dependencies, and thus the DbDependentTestBase is no longer needed
and was also removed.

Change-Id: I5abfbd31bd5608d3816f88e610faa98fb8fd4aa5
Signed-off-by: Allon Mureinik <amureini@redhat.com>