core: VdsStaticDao in GlusterVolumeDaoTest
GlusterVolumeDaoTest uses VdsStaticDao to retrieve a host from the
database, creating a dependency on another DAO, breaking its
"unitness".

This patch removes that faulty dependency, and uses arbitrary values
that are just as good for the test.
As a side bonus, removing this database operation also speeds up the
test.

Change-Id: I6070741bbaf39b2df665002876810a27c583f1a0
Signed-off-by: Allon Mureinik <amureini@redhat.com>