core: fix SPAPR_VSCSI controller and disks issues for ppc arch
Fix few bugs for supporting SPAPR_VSCSI disks (and cd-rom) in case of ppc
vm architecture by adding the following:
1. Create an unmanaged scsi controller device with index set to 0, to be
used by scsi cd-rom and SPAPR_VSCSI disks. This was done in 4.1 and is
now supported by engine xml as well.

2. Fix bugs related to default index in case of a SPAPR_VSCSI disk
by fixing the VmInfoBuildUtils.getDefaultVirtioScsiIndex() method
to consider the disk's interface while creating the vmDeviceSpaprVscsiUnitMap.
Before this fix, the default index was always set to VirtIO_SCSI, regardless
to actual disk interface and vm's architecture. That caused a NPE to be
raised in case of running a VM with SPAPR_VSCSI disk.

Change-Id: Ie2936689d2edfe929c9cd231a71e2623c4b2b2f1
Bug-Url: https://bugzilla.redhat.com/1565036
Signed-off-by: Sharon Gratch <sgratch@redhat.com>