core: Propagate MTU to VMs
The MTU of the logical network is propagated to libvirt's
definition of the VM [1], which will be propagated to
libvirt's tap devices, QEMU's command line and VirtIO drivers
inside the guest.

The MTU is respected by QEMU for rhel7.4.0 machine type and
later. It is ignored by QEMU with rhel7.3.0 machine type and
earlier.

Setting the MTU via libvirt will work only on hosts with
RHEL 7.4 or later. For hosts with RHEL 7.4 or later
libvirt's XML is produced by engine, which will contain the
definition. On hosts running older software, VDSM will create
libvirt's XML without the MTU definition. Setting the MTU via
libvirt works only on VMs with machine type rhel7.4.0 or
later, so setting the MTU via libvirt will work transparent to
the user if the default machine type in oVirt is rhel7.4.0 or
later.

[1] Documentation of libvirt's network XML format
https://libvirt.org/formatnetwork.html#elementsConnect

Change-Id: I4d282b28e2f0d1f4696ace1482d9b5644ce2a93a
Bug-Url: https://bugzilla.redhat.com/1451342
Signed-off-by: Dominik Holler <dholler@redhat.com>