core: query ova files from a directory
The current UI for import-VMs suggests that the path to query contains
a bunch of VMs because the result is inserted into a list that a subset
of it can be selected for import. This patch allows to provide a folder
as an input for import-VMs-from-OVA to comply with that. All OVA files
in that folder are then listed. The user can still provide a path to
a specific OVA though.

The changes are:
1. The script that queries the target path on the host is changed to
search all OVA files if the input is a folder.
2. The back-end parses new form of output of the query script and
converts it to a map of VM to its OVA filename.
3. The front-end is adjusted to receive that new output from the back-
end and convert it to a map of VM name to its OVA filename. Later, when
creating the parameters for the import command, this map is used to
construct the path to the OVA file.

Change-Id: Iae233016e2df8595b2628f5779ecf9dfc27a4068
Bug-Url: https://bugzilla.redhat.com/1568755
Signed-off-by: Arik Hadas <ahadas@redhat.com>