webadmin: Disable changing the name of external networks
If the name of the network entity representing an external
network is changed, this change is not propagated to the
external provider and the next synchronization will
overwrite the changed name with the original external one.

For this reason, webadmin prevents the user to change the name
of external networks.
The renaming is not blocked in the backend, because the
renaming functionality is still required for the
UpdateNetworkCommand.

Change-Id: I8e758b3aacf14c35e0f1fb581d6587402e79d0fa
Bug-Url: https://bugzilla.redhat.com/1583486
(cherry picked from commit 0f431d73cd9a52b6e430a9f211b9fdf47186ef5a)
Signed-off-by: Dominik Holler <dholler@redhat.com>