backend, packing: Add default MTU for tunnelled networks
The MTU of external networks might be smaller than the MTU
of the host networks, because the Ethernet frames of the
external networks might be transported by tunnels which adds
some bytes encapsulation overhead.

This change adds a new value to configuration to configure
the MTU of tunnelled networks the same way like the MTU of
non-tunnelled networks.

This implementation has the following limitations:
- It is assumed that traffic of external networks with
  pyhsnet is transported without tunnels.
- Custom physnets are ignored.
- The MTU of the physnet is ignored. This means that the
  default MTU of an external networks could be bigger than
  the MTU of the related physnet.

Change-Id: I7737f7d8b0afca8855203fb9609d474fc0d2211c
Bug-Url: https://bugzilla.redhat.com/1451342
Signed-off-by: Dominik Holler <dholler@redhat.com>