core: do not persist VmDevice#bootOrder
* Now that we don't rely on the persisted value of 'bootOrder' of VmDevice
(since the boot order is being computed on each run according to the
VM's default boot sequence or the boot sequence that is specified on
run-once flow, see I20b6ad), it can be removed from the database.

* On regular run-VM we used to rely on the boot order that is set on each
device, while on run-once we computed the boot order according to the
given boot sequence (if it is different than the default boot sequence
that is defined for the VM, otherwise it behaved as regular run).

* With this patch we always compute the boot order, not only on
run-once flow (on regular run it is done according to the VM's default
boot sequence). That way, we don't rely on the boot order that is set
on each device, so it enables us to remove this field and get rid of
the duplicated data whose management tends to complicate addition,
update and removal of bootable devices.

This patch is a squash of the two patches together.

Change-Id: Ibcabcc281698fa51b7b8f403010b8d57fbf9a574
Bug-Url: https://bugzilla.redhat.com/1568075
Signed-off-by: Arik Hadas <ahadas@redhat.com>
Signed-off-by: Daniel Erez <derez@redhat.com>