engine: Enforce r-curly checkstyle
The de-facto style of the oVirt Engine project is to have a right
curly followed directly by the next keyword instead of having that
keyword on the following line.

E.g., in an if statement:

    if (someCondition) {
        // Do something
    } else if (someOtherConditon) {
        // Do something else
    } else {
        // Do something entirely different
    }

Or in a try-catch-finally statement:

    try {
        // Something
    } catch (SomeException e) {
        // Handle exception
    } finally {
        // Cleanup
    }

This patch enforces this de-facto standard that is upheld in most of
the project with a checkstyle check, and fixes the [relatively] few
places that use a different convention to use this proper style.

Change-Id: I593879246b99383ef175ba417337f59d9af3086a
Signed-off-by: Allon Mureinik <amureini@redhat.com>