db: skip dropping extension functions
When extension functions are installed on the public schema that is used
also by the engine database, the upgrade procedure generates "drop
function" command for all installed functions in the public schema in
order to update changes, but in case of extension functions, those drop
commands will fail since the function depends on an installed extension
and can not be removed without removing the extension first.

This patch fixes the method that generates all those "drop function"
commands to skip extension functions.

Change-Id: I76a25d0677f2ebc64510bc905b509b7e939895ae
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1570988
Signed-off-by: emesika <emesika@redhat.com>