webadmin: Remove AdvancedViewColumns
AdvancedViewColumns is a second top-level class defined in
MainEventView.java, which doesn't align with neither the project's
conventions or Java's best practices.

Moreover, that class does nothing but holding a series of constants
that are just used in MainEventView, making it kind of pointless.

This patch inlines those constants in to MainEventView and removes
that redundant class.

Change-Id: I3520feffa9236f2ade08e88812fec4bea5280050
Signed-off-by: Allon Mureinik <amureini@redhat.com>