engine: Enforce one top level class per file
Java's best practices, as well as the oVirt Engine's coding style
advocate for a single top level class per .java file.

This patch fixes the few places where this rule is broken - either by
inlining the second class as an inner class of the first (where
possible) or extracting it to its own file (when there's no other
choice), and adds the OneTopLevelClass check in the project's
checkstyle configuration to make sure this mistake doesn't happen
again.

Change-Id: I69d64cdb479ab135bf414c7ec70a2b359c2e8889
Signed-off-by: Allon Mureinik <amureini@redhat.com>