Removing 'dns_resolver_configuration_id' from 'vds_dynamic' table
The id of the host will be used as the id of the
dns_resolver_configuration.

This change is done for two reasons -
1. There is one to one relationship between the host and its
dnsConfiguration.
A separate id for the dnsConfiguration is redundant.

2. In the flow of persisting the network topology after getting the host
capabilieties, vds_dynamic may be updated in two places-
  a. vds_dynamic.dns_resolver_configuration_id
  b. vds_dymanic.status (if a required network was removed from the host,
     the status of the host is changed to non-operational).

The two updates are done in separate transactions. The first
transaction is suspended while the second one is running. But since the
first one updated the vds_dynamic table the second one is blocked updating
the table.

Since the "vds_dynamic.dns_resolver_configuration_id" was redundant from
the first place, the easiest way to solve the bug was removing it.
Now it is gone, the first transaction is not updating the vds_dynamic table
so the second one has no problem to do the update.

Change-Id: I2bbc41301c718699a2efd89eeb9f8574d20335a6
Bug-Url: https://bugzilla.redhat.com/1570388
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>