core: ProcessOvfUpdateForStoragePoolCommandTest lists
The empty (default) return value of a list in a mock is an empty
list, so there's no need to explicitly mock such returns values. At
best, such mocking is just useless and at worst it's distracting and
draws the reader's attention to it.

This patch removes those redundant mocks from
ProcessOvfUpdateForStoragePoolCommandTest.

Change-Id: I063f47d505233bb5b52c141cf49fe7db053ddd02
Signed-off-by: Allon Mureinik <amureini@redhat.com>