core: StorageDomainValidatorTest lists
The empty (default) return value of a list in a mock is an empty
list, so there's no need to explicitly mock such returns values. At
best, such mocking is just useless and at worst it's distracting and
draws the reader's attention to it.

This patch removes those redundant mocks from
StorageDomainValidatorTest.

Change-Id: I0a5322e55ffc7e20b3dc061b4cd24fcba84f3384
Signed-off-by: Allon Mureinik <amureini@redhat.com>