webadmin: Adaptive list widget width in host dialog, affinity model tab
Previously, in the affinity model tab, in the host dialog, the list widget had a constant width.
When the button next to it has more text (for example in German),
it will hide the drop-down button of the list widget.

This patch makes the width of the list widget adaptive,
so it is visible even if the button is larger.

Change-Id: I5f16405c8cf45e6453d9241cf7075b6326bcd70b
Bug-Url: https://bugzilla.redhat.com/1542082
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>