core: filter loopback nics
We used to filter loopback NICs in ovirt-guest-agent. This capability
is missing in qemu-guest-agent. As a temporary solution, we now filter
them in the engine. This patch adds a regular expression that should
match loopbacks on various operating systems to the blacklist of guest
NIC names. Worse case, the user can modify this regular expression
if needed.

Change-Id: Ibab64028cb12cba0dfa05273618789116210c3be
Bug-Url: https://bugzilla.redhat.com/1571154
Signed-off-by: Arik Hadas <ahadas@redhat.com>