core: Handling unreachable agent in fencing flow
This patch adds a status request when agents are selected, adding the
possibility to pass to the next agent as soon as possible if the current
agent is unreachable from any reason.

This makes the other check that verifies that we are not trying to stop
a stopped machine or to start a started machine redundant.

The corresponding tests were moved from FenceVdsVDSCommandTest to
HostFenceActionExecutorTest and FenceVdsVDSCommandTest was removed.

Change-Id: I74377e704826fb9fa2992d13c0e4aea3c3e710c4
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1572148
Signed-off-by: emesika <emesika@redhat.com>