core: set vm as not-responsive only when guest cpu is running
When an attempt to gather VM statistics fails, VDSM reports negative
'monitorResponse' to indicate that the QEMU process does not function
properly. Previously, the engine always interpreted such a report to
a VM status of type NotResponsive. However, that was not necessarily
the right thing to do when the QEMU process was not supposed to be
functional, e.g., in case the VM is paused.

This patch changes the treatment of negavie monitorReponse reports.
Now they are mapped to NotReponsive VM status only when we know that
the VM is supposed to be functional.

Change-Id: Iba6bfe199c2c99203c9432058eb4b88b81b900b1
Bug-Url: https://bugzilla.redhat.com/1481022
Signed-off-by: Arik Hadas <ahadas@redhat.com>