restapi: UsageFinderTest assertions
Catching an exception in a test and then explicitly failing the test
is both cumbersome and misleading to the programmer trying to debug a
potential failure as it hides the actual cause of the failure.

This patch cleans up UsageFinderTest by allowing the exception to be
thrown upwards and the test to error-out instead of boiler-plate
catching it.

Change-Id: Ie29f158208fa783827d5b242515e94ae90412df8
Signed-off-by: Allon Mureinik <amureini@redhat.com>