engine: Don't extend org.junit.Assert
JUnit's best practice recommends to use Assert's static methods (or
even statically import them) and not extend Assert directly.
While this is currently only a recommendation, it will become
critical when we want to move to JUnit Jupiter, as Assertions cannot
be extended.

This patch follows JUnit's best practices, and replaces extending
org.junit.Assert with direct static imports of the relevant methods.

Change-Id: I2edf0825ab8868677a1a9addb4ec6b93c0975582
Signed-off-by: Allon Mureinik <amureini@redhat.com>