backend: forbid paused states for download flows
Today, an image download can be paused via the UI and the API.
The paused state only means that the engine won't monitor the transfer.
The download actually continues because it depends on the client, and it
can be canceled by the user if he wishes to stop it.

Therefore, this patch changes each status update of a download flow from
paused to canceled.

Change-Id: Ic72e4758a986dd25b644b350af89f09b6d894887
Related-To: https://bugzilla.redhat.com/1557770
Signed-off-by: Idan Shaby <ishaby@redhat.com>