core: use snapshot VM configuration when creating a template from snapshot
Currently, when creating a VM from a snapshot, the active VM configuration
was used when the command CreateAllTemplateDisksCommand  creates
the template disks.

The use of the active VM configuration caused a mismatch between the
disks that should be created and an exception was thrown when the
snapshot contains fewer disks than the active VM snapshot.

This patch fixes this behavior by using the snapshot VM configuration
instead of the active snapshot VM configuration

Change-Id: Ia50f492e43658fdc537fc68f4eeda8d5a6ffe04a
Bug-Url: https://bugzilla.redhat.com/1623818
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>