core: StorageDomainValidatorFreeSpaceTest snapshots
Having a DiskImage as a snapshot of itself creates a recursive data
structure where hashCode() fails with a StackOverflowError. While
this isn't currently a problem in this test, it may hinder future
refactoring efforts later on.

This patch uses a clone of the DiskImage as the snapshot in order to
make the data structure valid again. Storage Domain IDs were added
since the copyOf method requires them, even though they are not used
by the test.

Change-Id: I641037f4ffb50ace43d6f4a3b70ae167c20a4e7c
Signed-off-by: Allon Mureinik <amureini@redhat.com>