core: disable single-qxl-pci if needed on all import vm flows
Previously, we disabled single-qxl-pci, if needed, only when
importing from an external provider, assuming that in other
import flows it is taken care by the OVF parser.

This patch moves it to ImportVmCommandBase so it will be done
for importing from all sources, and in particular when importing
a configuration that is generated by virt-v2v, and it is now
done in the 'addVmStatic' method, which is a better place for
this logic then the command's 'init' method.

Change-Id: Id98ab1fadfc0c871663843c241562ff1f3f00b6d
Bug-Url: https://bugzilla.redhat.com/1573462
Signed-off-by: Arik Hadas <ahadas@redhat.com>