core: Remove unreachable code in MoveMacsTest
The ExpectedException rule does not prevent an exception from being
thrown - it only changes the test's behavior to only pass if the
given exception is thrown.

All the code in
testMigrateMacsToAnotherMacPoolWithUnsuccessfulDuplicityCheck after
the call to underTest will never be executed, as this call throws an
exception, and is probably only there as a result of a copy-paste
error.
This patch removes the unreachable code in order to clean up the test
and prevent confusion.

Change-Id: Ieaada177d0c80c52835b596929e8e032d4a04cd8
Signed-off-by: Allon Mureinik <amureini@redhat.com>