core: vmpools: keep a list of prestarted VMs before WaitForLaunch
After VmPoolMonitor prestarts some VMs, it takes time for them to create
a stateless snapshot and to switch to WaitForLaunch state. On the next
cycle VmPoolMonitor does not identify the VMs that are still Down and
have not created a stateless snapshot yet. It sees low number of
prestarted VMs and tries to run more VMs. Short VmPoolMonitor cycle
increases a probability of such behaviour.

To resolve this issue, we keep a list of VMs that were started by
VmPoolMonitor but not reached yet the WaitForLaunch state. These VMs
will be taken into account when calculating the number of prestarted
VMs.

Change-Id: I61f8ec24844136ed094b402df9ba8ffdbd8b7c2f
Bug-Url: https://bugzilla.redhat.com/1576752
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>