core: Extract MockConfigDescriptor
MockConfigDescriptor is essentially a POJO that just holds
information required to mock a configuration value.

This patch extracts it to be a top level class, and renames the
"mockConfig" method to a more standard "of".
While this doesn't have any functional benefit, it will help
transitioning to JUnit Jupiter, which will no longer support @Rules.

Change-Id: Ibb3e2fb45e7d6baef9819d7caa8e65e2a573f391
Signed-off-by: Allon Mureinik <amureini@redhat.com>