core: UpdateVmCommandTest config mocking
UpdateVmCommandTest uses MockConfigRule as a static @ClassRule, but
allows the instance method testValidName to modify the mocked
configuration. Since it's a @ClassRule, the changes this method makes
will persist to other tests and thus break their atomicity.
In fact, a lot of the classes tests pass by sheer luck, due to the
added configuration mocking this test performs, and will fail if run
alone.

This patch fixes the test class so all the configuration mocking is
done in one place (the rule), except for the HE specific mocking,
which although may technically be wrong, does not affect the other
tests.

Change-Id: If2954b98790ca70bc17ce5eebb925bb116fdab8c
Signed-off-by: Allon Mureinik <amureini@redhat.com>