core: UpdateVmDiskCommandTest config mocking
UpdateVmDiskCommandTest uses MockConfigRule as a static @ClassRule,
but allows the instance method testValidName to modify the mocked
configuration. Since it's a @ClassRule, the changes this method makes
will persist to other tests and thus break their atomicity.
In fact, a lot of the classes tests pass by sheer luck, due to the
added configuration mocking this test performs, and will fail if run
alone.

This patch fixes the test class so all the configuration mocking is
done in one place (the rule).

Change-Id: I3c4239f49f3993f3cc3e89cdf9e5ec56be859f4f
Signed-off-by: Allon Mureinik <amureini@redhat.com>