core: Extract PolicyUnitImpl#getMaxSchedulerWeight()
PolicyUnitImpl#MaxSchedulerWeight depends on a ConfigValue, which may
or may not be available when the class is loaded, which may cause the
class loading to fail. The fact that many tests that should not care
about this configuration value are forced to mock it shows this
potential issue.

Extracting it to a method protects against such potential critical
bugs.

Change-Id: Ibdfa8b0f70311e236c307e8651e32f3bc2c8c99b
Signed-off-by: Allon Mureinik <amureini@redhat.com>