core: do not remove vm async running command on power-off
Now that we do not change the status of the VM when doing an
ungraceful destroy, we should not remove the VM async running
command too soon as we may clear migrating_to_vds during the on-
going migration and that may result in doing hand-over to a
'null' destination. Instead, the VM async running command will
be removed by the monitoring thread when the VM disappears.

Change-Id: I905632ac394b87a306d308d7ddefcba27acbba0d
Bug-Url: https://bugzilla.redhat.com/1558709
Signed-off-by: Arik Hadas <ahadas@redhat.com>