core: fix power-off migrating vm
Powering off a migrating VM is comprised of two phases. First, the VM
is destroyed on the destination host. Second, it is destroyed on the
source host. Doing that vice versa may result in having the VM paused
on the destination host.

We now lock the VM during this process, preventing it from being
monitored at that time. This is required in order to prevent races
that may end up in trying to rerun the migration when detecting that
the VM went down on the destination host.

Change-Id: I795aff4002d1f257652ca3b44da32b62e30d2c5e
Bug-Url: https://bugzilla.redhat.com/1558709
Signed-off-by: Arik Hadas <ahadas@redhat.com>