core: IsoDomainListSynchronizer#getRegexToolPattern
The REGEX_PATTERN_TOOL is a static member (i.e., it's initialized
when the class is loaded), but it relies on a configuration value
fetched from the database, which may or may not be available when the
class is loaded, leading to hard-to-detect bugs.

Since this value is cached anyway, and getting it is negligible
compared to compiling the pattern it's used in (not to mention
retrieving ISOs from VDSM), wrapping it inside an on-demand getter
seems like the right way to go.

The fact that moving this dependency on the Config out of the class
loading allows removing the empty MockConfigRule from
RunVmCommandTest serves as a further indicator that this was the
right decision.

Change-Id: I0d3a52afb16591634c984cfa22e692fbee9f1088
Signed-off-by: Allon Mureinik <amureini@redhat.com>