core: OvfManagerTest#testRemoveImageFromSnapshotConfiguration
RemoveImageCommandTest does not actually test the RemoveImageCommand,
but only the OvfManager. As such, it makes little to keep it, and
confuse the person who's supposed to maintain it.

This patch moves the only test in RemoveImageCommandTest,
testRemoveImageFromSnapshotConfiguration to OvfManagerTest, where it
seems to better belong, along with the helper methods and mocked
members required for it.

Change-Id: I3ed13d0b76647156bc6033a75c9cc7ae0d838cf9
Signed-off-by: Allon Mureinik <amureini@redhat.com>