webadmin: show snapshot disks while creating a template from a snapshot
Currently, when the user selects to create a template from snapshot via the
UI, the disks that appear in the pop-up window belongs to the active VM
instead of the VM selected snapshot.

This patch will fix this behavior by retrieving the snapshots disks and
present them in case of creating a template from a snapshot.

Change-Id: Ice283c75c261c1696ab450776a5cdbfb84c0ebb2
Bug-Url: https://bugzilla.redhat.com/1575586
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>