core: Bump Spring dependency
Upgrade the oVirt Engine's Spring dependency to the latest 5.0.4.
Alongside some general security and performance improvements, the two
main motivations for this upgrade are the native support for Java 8
and the facilitation of a migration to JUnit Jupiter.

Note that Spring JDBC 5 is slightly stricter than Spring JDBC 4, and
no longer allows null MapSqlParameterSources. Instead, in the few
places that used to pass nulls there, empty instances are used.
Findbugs exclusions were also added to the flows where Spring added
correctness annotations, but the code was tested to work correctly.

Change-Id: Ia0b1f0f8ef72efaa9cc07aec73dcbea83976b4ce
Signed-off-by: Allon Mureinik <amureini@redhat.com>