core: VdsBrokerObjectsBuilderTest maps
VdsBrokerObjectsBuilderTest uses Guava's ImmutableMap, which comes
from a dependency that is never declared, and just happens to be in
the classpath by some indirect dependency that may disappear one day.

This patch removes those usages and replaces them with the standard
Collections.singletonMap call used in other places throughout the
codebase, including in this very class.

Change-Id: I3016c4c8df242f31f23d33b0c349014ec07e1298
Signed-off-by: Allon Mureinik <amureini@redhat.com>