core: ignore noVM when monitoring destroys migrating vm
When VMs monitoring decides to destroy a migrating VM, an error of
type noVM that we get back from the destination host should be ignored
as it most typically does not indicate of a real error. In particular,
this happens when the user asks to power-off migrating VM.

Previously, we ignored noVM error when destroying a migrating VM only
when that error was returned from the source host. Now, we ignore it
also when it is returned from the destination host.

Change-Id: I04730156c8867e5ce68d56b28c54d1010928a7f9
Bug-Url: https://bugzilla.redhat.com/1558709
Signed-off-by: Arik Hadas <ahadas@redhat.com>