restapi: Fix Affinity Label Links
When Affinity-Labels are displayed in the context of a VM:
.../api/vms/xxx/affinitylabels

Their links to the vms, hosts are not displayed.

In the context of this bug it was revelaed that getting a
single Affinity-Label under a VM does not work at all.

This patch fixes both problems

Bug-Url: https://bugzilla.redhat.com/1575596
Change-Id: I65d675629ee2480840b11638f6d4d2fa6b1fc4cf
Signed-off-by: Ori Liel <oliel@redhat.com>
(cherry picked from commit 8342797c6147227bf5d0745ec1c38f6e89bb0e5a)