core: import vm template from configuration
Adjusting ImportVmTemplateFromConfiguration to the use-case of importing a
template using virt-v2v - first, the disks are being uploaded to the data
domain and then this command is called with template configuration and
should add the VmTemplate entity and attach the uploaded disk to that entity.

The required changes are:
1. Skipping validations that are irrelevant to this flow.
2. Attaching the disks that appear in the configuration.

The added parts are similar to those in ImportVmFromConfiguration for
supporting import VM using virt-v2v.

Change-Id: Ia564eeb7779e147b5ba287bcaaf17333e5632cbf
Bug-Url: https://bugzilla.redhat.com/1526032
Signed-off-by: Arik Hadas <ahadas@redhat.com>