rest-api: use vm-init when adding vm/template from configuration
The configuration for adding vm/template from configuration is specified
in the initialization part of the VM. This creates the following
problem: when the initialization contains only the configuration with no
vm-init then the vm-init is read from the configuration and then
overridden with empty vm- init (because the initialization part contains
only the configuration but no vm-init data).

This patch clears the vm-init from the initialization part immediately
after reading the configuration so the vm-init data that is read from
the configuration will not be overridden. This is done for both VMs and
template.

Change-Id: Iffa99f06e5a5ae8dd07756be1de135dff0dad101
Bug-Url: https://bugzilla.redhat.com/1560553
Signed-off-by: Arik Hadas <ahadas@redhat.com>